+++ ovos_config/config.py
@@ -27,15 +27,6 @@ from ovos_utils.json_helper import flattened_delete, merge_dict
 from ovos_utils.log import LOG
 
 
-def _log_old_location_deprecation(old_user_config=OLD_USER_CONFIG):
-    LOG.warning(" ===============================================")
-    LOG.warning(" ==             DEPRECATION WARNING           ==")
-    LOG.warning(" ===============================================")
-    LOG.warning(f" You still have a config file at {old_user_config}")
-    LOG.warning(" Note that this location is deprecated and will" +
-                " not be used in the future")
-    LOG.warning(" Please move it to " + get_xdg_config_save_path())
-
 
 class Configuration(dict):
     """Namespace for operations on the configuration singleton."""
@@ -48,10 +39,6 @@ class Configuration(dict):
     # This includes both the user config and
     # /etc/xdg/mycroft/mycroft.conf
     xdg_configs = [LocalConf(p) for p in get_xdg_config_locations()]
-    _old_user = LocalConf(OLD_USER_CONFIG)
-    # deprecation warning
-    if isfile(OLD_USER_CONFIG):
-        _log_old_location_deprecation(OLD_USER_CONFIG)
     _watchdog = None
     _callbacks = []
 
@@ -65,7 +52,7 @@ class Configuration(dict):
         # sync with other processes connected to bus
         if Configuration.bus:
             # imported from ovos_utils to allow FakeMessage if ovos-bus-client is missing
-            from ovos_utils.messagebus import Message
+            from ovos_utils.fakebus import Message
             Configuration.bus.emit(Message("configuration.patch",
                                            {"config": {key: value}}))
 
@@ -187,9 +174,6 @@ class Configuration(dict):
         if not skip_remote:
             configs.insert(1, Configuration.remote)
         if not skip_user:
-            # deprecation warning
-            if isfile(OLD_USER_CONFIG):
-                configs.append(Configuration._old_user)
             configs += Configuration.xdg_configs
 
         # runtime patches by skills / bus events